+++ src/commands/new_stacked_view_view_model/generate_stacked_view_view_model_code.ts
@@ -1,10 +1,10 @@
 import { existsSync } from "fs";
-import { createDirectory } from "../../utils";
 import {
   createStackedViewModelTemplate,
   createStackedViewTemplate,
 } from "./create_stacked_view_view_model_templates";
-import { Utils } from "../../utils";
+import { Utils, promptForOverrideDirectory } from "../../utils";
+import { FileSystemManager } from "../../utils/file_system_manager";
 
 export async function generateStackedViewViewModelCode(
   stackedName: string,
@@ -14,13 +14,32 @@ export async function generateStackedViewViewModelCode(
   const snakeCaseStackedName = Utils.convertToSnakeCase(
     stackedName.toLowerCase()
   );
+  const pascalCaseStackedName = Utils.convertToPascalCase(
+    stackedName.toLowerCase()
+  );
+
   const stackedDirectoryPath = `${targetDirectory}/${snakeCaseStackedName}`;
+
   if (!existsSync(stackedDirectoryPath)) {
-    await createDirectory(stackedDirectoryPath);
+    await FileSystemManager.createFolder(stackedDirectoryPath);
+    await Promise.all([
+      createStackedViewModelTemplate(stackedName, targetDirectory),
+      createStackedViewTemplate(stackedName, targetDirectory, useReactive),
+    ]);
+    Utils.showInformationMessage(
+      `Successfully Generated ${pascalCaseStackedName} Stacked View-ViewModel`
+    );
+    return;
   }
 
-  await Promise.all([
-    createStackedViewModelTemplate(stackedName, targetDirectory),
-    createStackedViewTemplate(stackedName, targetDirectory, useReactive),
-  ]);
+  if (existsSync(stackedDirectoryPath)) {
+    Utils.showInformationMessage(
+      `
+      \"${pascalCaseStackedName}\" Directory Already Exists!
+      \n
+      Did Not Create Stacked View-ViewModel.
+      `
+    );
+    return;
+  }
 }